fix integer divide rounding issue, fix compiler signed/unsigned warnings, and defined-but-not-used warning caused by non-const declaration of a const

+++ src/main.cpp
@@ -1012,7 +1012,7 @@ bool AcceptToMemoryPool(CTxMemPool& pool, CValidationState &state, const CTransa
     if (pfMissingInputs)
         *pfMissingInputs = false;
 
-    int64_t maxBlockSize = Params().GetConsensus().MaxBlockSize(GetAdjustedTime(), sizeForkTime.load());
+    uint64_t maxBlockSize = Params().GetConsensus().MaxBlockSize(GetAdjustedTime(), sizeForkTime.load());
 
     if (!CheckTransaction(tx, state, maxBlockSize))
         return error("AcceptToMemoryPool: CheckTransaction failed");
@@ -1156,13 +1156,13 @@ bool AcceptToMemoryPool(CTxMemPool& pool, CValidationState &state, const CTransa
 	//   Some free transactions will still get through based on -limitfreerelay
         float feeCutoff;
         uint8_t nFreeLimit;
-        if (poolBytes < maxBlockSize){
+        if (poolBytes < maxBlockSize){  // Nothing to do because all txns can fit in a block.
             feeCutoff = 0;
             nFreeLimit = nLimitFreeRelay;
         }
         else if(poolBytes < maxBlockSize * MAX_BLOCK_SIZE_MULTIPLYER){
             // Gradually choke off what is considered a free transaction
-            feeCutoff = (minRelayTxFee/1000) * (poolBytes-maxBlockSize)/(maxBlockSize * (MAX_BLOCK_SIZE_MULTIPLYER-1));
+            feeCutoff = (minRelayTxFee/1000) * (poolBytes-maxBlockSize) / (maxBlockSize * (MAX_BLOCK_SIZE_MULTIPLYER-1));
 
             // Gradually choke off the nFreeLimit as well but leave at least minFreeLimit
             // So that some free transactions can still get through
@@ -1175,7 +1175,7 @@ bool AcceptToMemoryPool(CTxMemPool& pool, CValidationState &state, const CTransa
             nFreeLimit = MIN_LIMIT_FREE_RELAY;
         }
 
-        double satoshiPerByte = nFees/nSize;
+        double satoshiPerByte = ((double)nFees)/nSize;
         LogPrint("mempool",
                  "MempoolBytes:%d  LimitFreeRelay:%d  FeeCutOff:%.4g  FeesSatoshiPerByte:%.4g  TxBytes:%d  TxFees:%d\n",
                   poolBytes,nFreeLimit,feeCutoff,satoshiPerByte, nSize, nFees);
